+++ loinc2hpo-core/src/main/java/org/monarchinitiative/loinc2hpo/fhir/FHIRLoincPanelFactory.java
@@ -1,0 +1,10 @@
+package org.monarchinitiative.loinc2hpo.fhir;
+
+import org.hl7.fhir.dstu3.model.Patient;
+import org.monarchinitiative.loinc2hpo.loinc.LoincId;
+
+public interface FHIRLoincPanelFactory {
+    FHIRLoincPanel createFhirLoincPanel(LoincId loincId);
+
+    FHIRLoincPanel createFhirLoincPanel(LoincId loincId, Patient patient);
+}