+++ loinc2hpo-core/src/test/java/org/monarchinitiative/loinc2hpo/fhir/FhirObservationRetrieverTest.java
@@ -8,6 +8,7 @@ import ca.uhn.fhir.rest.api.Constants;
 import ca.uhn.fhir.rest.api.MethodOutcome;
 import ca.uhn.fhir.rest.client.api.IGenericClient;
 import org.hl7.fhir.dstu3.model.*;
+import org.junit.Ignore;
 import org.junit.Test;
 
 
@@ -32,6 +33,7 @@ public class FhirObservationRetrieverTest {
     }
 
     @Test
+    @Ignore
     public void testUploadBundle() {
         // Create a patient object
         Patient patient = new Patient();