+++ loinc2hpo-core/src/test/java/org/monarchinitiative/loinc2hpo/fhir/ObservationAnalysisFromInterpretationTest.java
@@ -9,8 +9,9 @@ import org.monarchinitiative.loinc2hpo.codesystems.Code;
 import org.monarchinitiative.loinc2hpo.exception.AmbiguousResultsFoundException;
 import org.monarchinitiative.loinc2hpo.exception.MalformedLoincCodeException;
 import org.monarchinitiative.loinc2hpo.loinc.*;
+import org.monarchinitiative.phenol.base.PhenolException;
 import org.monarchinitiative.phenol.formats.hpo.HpoOntology;
-import org.monarchinitiative.phenol.io.obo.hpo.HpoOboParser;
+import org.monarchinitiative.phenol.io.obo.hpo.HpOboParser;
 import org.monarchinitiative.phenol.ontology.data.Term;
 import org.monarchinitiative.phenol.ontology.data.TermId;
 
@@ -41,11 +42,11 @@ public class ObservationAnalysisFromInterpretationTest {
         observations[1] = observation2;
 
         String hpo_obo = FhirObservationAnalyzerTest.class.getClassLoader().getResource("obo/hp.obo").getPath();
-        HpoOboParser hpoOboParser = new HpoOboParser(new File(hpo_obo));
+        HpOboParser hpoOboParser = new HpOboParser(new File(hpo_obo));
         HpoOntology hpo = null;
         try {
             hpo = hpoOboParser.parse();
-        } catch (IOException e) {
+        } catch (PhenolException e) {
             e.printStackTrace();
         }
         ImmutableMap.Builder<String,Term> termmap = new ImmutableMap.Builder<>();