+++ loinc2hpogui/src/main/java/org/monarchinitiative/loinc2hpo/controller/AnnotateTabController.java
@@ -5,7 +5,6 @@ import com.google.common.collect.ImmutableMap;
 import com.google.inject.Inject;
 import com.google.inject.Injector;
 import com.google.inject.Singleton;
-import javafx.animation.FadeTransition;
 import javafx.application.Platform;
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.ReadOnlyStringWrapper;
@@ -18,19 +17,21 @@ import javafx.collections.ObservableList;
 import javafx.concurrent.Task;
 import javafx.event.ActionEvent;
 import javafx.event.Event;
-import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
-import javafx.scene.Node;
+import javafx.geometry.Insets;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.scene.input.*;
+import javafx.scene.layout.Background;
+import javafx.scene.layout.BackgroundFill;
+import javafx.scene.layout.GridPane;
+import javafx.scene.layout.VBox;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Circle;
 import javafx.stage.*;
 import javafx.util.Callback;
-import javafx.util.Duration;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.monarchinitiative.loinc2hpo.codesystems.Code;
@@ -50,16 +51,16 @@ import org.monarchinitiative.loinc2hpo.io.WriteToFile;
 import org.monarchinitiative.loinc2hpo.loinc.*;
 import org.monarchinitiative.loinc2hpo.model.AdvantagedAnnotationTableComponent;
 import org.monarchinitiative.loinc2hpo.model.Model;
-import org.monarchinitiative.loinc2hpo.util.HPO_Class_Found;
-import org.monarchinitiative.loinc2hpo.util.LoincLongNameComponents;
-import org.monarchinitiative.loinc2hpo.util.LoincLongNameParser;
-import org.monarchinitiative.loinc2hpo.util.SparqlQuery;
+import org.monarchinitiative.loinc2hpo.util.*;
 import org.monarchinitiative.phenol.ontology.data.Term;
 
 
 import java.io.*;
 import java.time.LocalDateTime;
 import java.util.*;
+import java.util.stream.Collectors;
+
+import org.monarchinitiative.loinc2hpo.gui.ColorUtils;
 
 
 @Singleton
@@ -120,6 +121,9 @@ public class AnnotateTabController {
     @FXML private TableColumn<LoincEntry, String> scaleTableColumn;
     @FXML private TableColumn<LoincEntry, String> systemTableColumn;
     @FXML private TableColumn<LoincEntry, String> nameTableColumn;
+    @FXML private CheckMenuItem loincTableEnableMultiSelection;
+    private LOINC2HpoAnnotationImpl toCopy;
+    @FXML private MenuItem pasteAnnotationButton;
 
 
     @FXML private Button modeButton;
@@ -156,13 +160,15 @@ public class AnnotateTabController {
     @FXML private Menu loincListsButton;
     @FXML private MenuItem backMenuItem;
     @FXML private MenuItem forwardMenuItem;
-    @FXML private Menu userCreatedLoincListsButton;
+    @FXML private Menu groupUngroup2LoincListButton;
     @FXML private Menu exportLoincListButton;
     @FXML private Menu importLoincGroupButton;
     final ObservableList<String> userCreatedLoincLists = FXCollections
             .observableArrayList();
     final private String LOINCWAITING4NEWHPO = "require_new_HPO_terms";
     final private String LOINCUNABLE2ANNOTATE = "unable_to_annotate";
+    final private String UNSPECIFIEDSPECIMEN = "unspecified_specimen";
+    final private String LOINC4QC = "test_for_QC";
 
     private BooleanProperty isPresentOrd = new SimpleBooleanProperty(false);
 
@@ -239,8 +245,7 @@ public class AnnotateTabController {
             }
         });
 
-        //if user creates a new Loinc group, add two menuitems for it, and specify the actions when those menuitems are
-        //clicked
+        //if user creates a new Loinc group, add two menuitems for it, and specify the actions when those menuitems are clicked
         userCreatedLoincLists.addListener(new ListChangeListener<String>() {
             @Override
             public void onChanged(Change<? extends String> c) {
@@ -255,7 +260,7 @@ public class AnnotateTabController {
                                 model.addUserCreatedLoincList(p, new LinkedHashSet<>());
 
                                 MenuItem newListMenuItem = new MenuItem(p);
-                                userCreatedLoincListsButton.getItems().add(newListMenuItem);
+                                groupUngroup2LoincListButton.getItems().add(newListMenuItem);
                                 newListMenuItem.setOnAction((event -> {
                                     logger.trace("action detected");
                                     if (loincTableView.getSelectionModel().getSelectedItem()!=null) {
@@ -372,11 +377,37 @@ public class AnnotateTabController {
         });
 
 
+        loincTableEnableMultiSelection.selectedProperty().addListener(new ChangeListener<Boolean>() {
+            @Override
+            public void changed(ObservableValue<? extends Boolean> observable, Boolean oldValue, Boolean newValue) {
+                if (observable != null) {
+                    if (newValue) {
+                        logger.trace("multi selection is enabled");
+                        loincTableView.getSelectionModel().setSelectionMode(SelectionMode.MULTIPLE);
+                    } else {
+                        logger.trace("multi selection is not enabled");
+                        loincTableView.getSelectionModel().setSelectionMode(SelectionMode.SINGLE);
+                    }
+                } else {
+                    return;
+                }
+            }
+        });
     }
 
     void defaultStartUp() {
         initLOINCtable(null);
         Platform.runLater(()->initHPOmodelButton(null));
+        List<String> DEFAULTGROUPS = Arrays.asList(LOINCWAITING4NEWHPO, LOINCUNABLE2ANNOTATE, UNSPECIFIEDSPECIMEN, LOINC4QC);
+        for(int i = 0; i < DEFAULTGROUPS.size(); i++) {
+            if (!model.getUserCreatedLoincListsColor().containsKey(DEFAULTGROUPS.get(i))) {
+                model.addOrUpdateUserCreatedLoincListColor(DEFAULTGROUPS.get(i), model.defaultColorList().get(i + 1));
+                logger.info(DEFAULTGROUPS.get(i) + "::::" + model.defaultColorList().get(i + 1));
+            }
+        }
+        logger.info("default colors in model:");
+        //model.defaultColorList().forEach(System.out::println);
+        logger.trace("default color for LOINC lists is set");
     }
 
     private void noLoincEntryAlert(){
@@ -745,7 +776,7 @@ public class AnnotateTabController {
                     .filter( loincEntry -> containedIn(query, loincEntry.getLongName()))
                     .forEach(loincEntry -> {
                         entrylist.add(loincEntry);
-                        logger.debug(loincEntry.getLOINC_Number() + " : " + loincEntry.getLongName());
+                        //logger.debug(loincEntry.getLOINC_Number() + " : " + loincEntry.getLongName());
                     });
                     //.forEach(loincEntry -> entryListInOrder.add(loincEntry));
         }
@@ -884,9 +915,78 @@ public class AnnotateTabController {
         }
         //model.addUserCreatedLoincList(nameOfList, new LinkedHashSet<>());
         userCreatedLoincLists.add(nameOfList);
-
+        Random rand = new Random();
+        double[] randColorValues = rand.doubles(3, 0, 1).toArray();
+        Color randColor = Color.color(randColorValues[0], randColorValues[1], randColorValues[2]);
+        model.addOrUpdateUserCreatedLoincListColor(nameOfList, ColorUtils.colorValue(randColor));
     }
 
+    @FXML
+    private void setLoincGroupColor(ActionEvent e) {
+        logger.trace("user wants to set the color of LOINC groups");
+        Stage window = new Stage();
+
+        VBox root = new VBox();
+        root.setSpacing(10);
+
+
+        ToolBar toolBar = new ToolBar();
+        final ComboBox<String> loincGroupCombo = new ComboBox();
+
+        loincGroupCombo.getItems().addAll(userCreatedLoincLists);
+
+        final ColorPicker colorPicker = new ColorPicker();
+        loincGroupCombo.getSelectionModel().selectedItemProperty().addListener(new ChangeListener<String>() {
+            @Override
+            public void changed(ObservableValue<? extends String> observable, String oldValue, String newValue) {
+                if (observable != null) {
+                    final String colorString = model.getUserCreatedLoincListsColor().get(newValue);
+                    if (colorString != null) {
+                        final Color color = Color.web(colorString);
+                        colorPicker.setValue(color);
+                    }
+
+                }
+            }
+        });
+        loincGroupCombo.getSelectionModel().select(0);
+
+        toolBar.getItems().addAll(loincGroupCombo, colorPicker);
+
+        GridPane gridPane = new GridPane();
+        gridPane.setHgap(10);
+        gridPane.setVgap(10);
+        gridPane.setPadding(new Insets(2));
+
+        Map<String, TextField> gridPaneColors = new HashMap<>();
+        for (int i = 0; i < userCreatedLoincLists.size(); i++) {
+            TextField name = new TextField(userCreatedLoincLists.get(i));
+            gridPane.add(name, 0, i);
+            TextField color = new TextField();
+            color.setBackground(new Background(new BackgroundFill(Color.web(model.getUserCreatedLoincListsColor().get(name.getText())), null, null)));
+            gridPane.add(color, 1, i);
+            gridPaneColors.put(name.getText(), color);
+            logger.trace("color" + color.getBackground().getFills().toString());
+        }
+
+
+        colorPicker.setOnAction(t -> {
+            model.addOrUpdateUserCreatedLoincListColor(loincGroupCombo.getSelectionModel().getSelectedItem(), colorPicker.getValue().toString());
+            logger.trace("new color: " + model.getUserCreatedLoincListsColor().get(loincGroupCombo.getSelectionModel().getSelectedItem()));
+            gridPaneColors.get(loincGroupCombo.getSelectionModel().getSelectedItem()).setBackground(new Background(new BackgroundFill(colorPicker.getValue(), null, null)));
+            changeColorLoincTableView();
+            model.writeSettings();
+        });
+
+        root.getChildren().addAll(toolBar, gridPane);
+        Scene scene = new Scene(root, 400, 400);
+        window.setScene(scene);
+        window.showAndWait();
+    }
+
+    /**
+     * This function is normally not useful because of MainController::openSession()
+     */
     private void initializeUserCreatedLoincListsIfNecessary(){
         //execute the functionalities only once in each secession
         if (!model.getUserCreatedLoincLists().isEmpty()) {
@@ -894,17 +994,21 @@ public class AnnotateTabController {
             return;
         }
         //by default, there will be two user created lists
+        //This is not scaling well. @TODO: consider other ways
+        //consider detecting existing lists by scanning the folder
         List<String> initialListNames = new ArrayList<>();
         initialListNames.add(LOINCWAITING4NEWHPO);
         initialListNames.add(LOINCUNABLE2ANNOTATE);
+        initialListNames.add(UNSPECIFIEDSPECIMEN);
+        initialListNames.add(LOINC4QC);
         userCreatedLoincLists.addAll(initialListNames);
         logger.trace("initializeUserCreatedLoincListsIfNecessary(): 2222");
         /*
         //create a menuitem for each and add to two menus; also create a list to record data
-        userCreatedLoincListsButton.getItems().clear();
+        groupUngroup2LoincListButton.getItems().clear();
         exportLoincListButton.getItems().clear();
         initialListNames.forEach(p -> {
-            userCreatedLoincListsButton.getItems().add(new MenuItem(p));
+            groupUngroup2LoincListButton.getItems().add(new MenuItem(p));
             exportLoincListButton.getItems().add(new MenuItem(p));
             model.addUserCreatedLoincList(p, new ArrayList<>());
         });
@@ -943,7 +1047,7 @@ public class AnnotateTabController {
         e.consume();
         logger.trace("context memu for loinc table requested");
         initializeMenuItemsForFilteredLists();
-        initializeUserCreatedLoincListsIfNecessary();
+        //initializeUserCreatedLoincListsIfNecessary(); //usually not run
         logger.trace("exit buildContextMenuForLoinc()");
     }
 
@@ -1053,7 +1157,7 @@ public class AnnotateTabController {
 
     @FXML private void handleCandidateHPODragged(MouseEvent e) {
 
-        System.out.println("Drag event detected");
+        //System.out.println("Drag event detected");
         Dragboard db = hpoListView.startDragAndDrop(TransferMode.ANY);
         ClipboardContent content = new ClipboardContent();
         Object selectedCell = hpoListView.getSelectionModel().getSelectedItem();
@@ -1135,6 +1239,101 @@ public class AnnotateTabController {
         return inverseChecker.isSelected();
     }
 
+    //ensure that at least one LOINC entry is selected
+    private boolean checkSelectedLoinc() {
+        switch (loincTableView.getSelectionModel().getSelectionMode()) {
+            case SINGLE:
+                if (loincTableView.getSelectionModel().getSelectedItem() != null &&
+                        loincTableView.getSelectionModel().getSelectedItem().getScale().equals(toCopy.getLoincScale().toString())) {
+                    return true;
+                }
+            case MULTIPLE:
+                if (loincTableView.getSelectionModel().getSelectedItems() != null &&
+                        !loincTableView.getSelectionModel().getSelectedItems().isEmpty()) {
+                    List<String> scaleTypes = loincTableView.getSelectionModel().getSelectedItems()
+                            .stream().map(LoincEntry::getScale)
+                            .distinct().collect(Collectors.toList());
+                    if (scaleTypes.size() == 1 && scaleTypes.get(0).equals(toCopy.getLoincScale().toString())) {
+                        return true;
+                    }
+                }
+            default:
+                return false;
+        }
+    }
+
+    @FXML private void copyAnnotation(ActionEvent event) {
+        logger.trace("copy loincAnnotation to other LOINCs");
+        LoincEntry selectedLoinc = loincTableView.getSelectionModel().getSelectedItem();
+        if (selectedLoinc == null) {
+            logger.error("Select a LOINC entry to copy");
+            return;
+        }
+        if (!model.getLoincAnnotationMap().containsKey(selectedLoinc.getLOINC_Number())) {
+            PopUps.showWarningDialog("Error Selection", "LOINC does not have annotation", "Select a LOINC code that has already been annotated");
+            logger.error("Annotation does not exist for " + selectedLoinc.getLOINC_Number());
+            return;
+        }
+        toCopy = model.getLoincAnnotationMap().get(selectedLoinc.getLOINC_Number());
+        loincTableEnableMultiSelection.setSelected(true);
+        //loincTableView.getSelectionModel().setSelectionMode(SelectionMode.MULTIPLE);
+    }
+
+    @FXML private void pasteAnnotation(ActionEvent e) {
+        logger.trace("paste annotation");
+        if (toCopy == null) {
+            PopUps.showWarningDialog("Error Selection", "Unspecified item to copy", "Select a LOINC code to copy annotation from in the LOINC table");
+            return;
+        }
+        if (!checkSelectedLoinc()) { //make sure there is at least one valid LOINC entry selection
+            PopUps.showWarningDialog("Error Selection", "Possible Errors:", "1) No LOINC entry is selected;\n2) >= 1 selected LOINC entry does match the scale of origin LOINC");
+            return;
+        }
+
+        if (!loincTableEnableMultiSelection.isSelected()) {
+            pasteAnnotationTo(loincTableView.getSelectionModel().getSelectedItem().getLOINC_Number());
+        } else {
+            loincTableView.getSelectionModel().getSelectedItems().stream()
+                    .forEach(loinc -> {
+                        logger.trace("copy to: " + loinc.getLongName());
+                        pasteAnnotationTo(loinc.getLOINC_Number());
+                    });
+        }
+
+        //refresh annotation tab
+        loinc2HpoAnnotationsTabController.refreshTable();
+        changeColorLoincTableView();
+        model.setSessionChanged(true);
+
+        //reset
+        toCopy = null;
+        //loincTableView.getSelectionModel().setSelectionMode(SelectionMode.SINGLE);
+        loincTableEnableMultiSelection.setSelected(false);
+
+    }
+
+    private void pasteAnnotationTo(LoincId loincId) {
+        if (model.getLoincAnnotationMap().containsKey(loincId)) {
+            logger.trace("Overwrite: " + loincId);
+        }
+        String comments = toCopy.getNote()==null ? "" : "@original comment: " + toCopy.getNote();
+        String copyInfo = String.format("copied from: %s %s", toCopy.getLoincId().toString(), comments);
+
+        LOINC2HpoAnnotationImpl.Builder builder = new LOINC2HpoAnnotationImpl.Builder()
+                .setLoincId(loincId)
+                .setLoincScale(toCopy.getLoincScale())
+                .setCreatedBy(toCopy.getCreatedBy())
+                .setCreatedOn(toCopy.getCreatedOn())
+                .setLastEditedBy(toCopy.getLastEditedBy())
+                .setLastEditedOn(toCopy.getLastEditedOn())
+                .setNote(copyInfo)
+                .setFlag(toCopy.getFlag())
+                .setVersion(toCopy.getVersion());
+        toCopy.getCandidateHpoTerms().entrySet().stream()
+                .forEach(entry -> builder.addAdvancedAnnotation(entry.getKey(), entry.getValue()));
+        model.addLoincTest(builder.build());
+
+    }
 
     @FXML private void createLoinc2HpoAnnotation(ActionEvent e) {
 
@@ -1143,6 +1342,7 @@ public class AnnotateTabController {
             PopUps.showInfoMessage("No loinc entry is selected. Try clicking \"Initialize Loinc Table\"", "No Loinc selection Error");
             return;
         }
+
         //specify loinc for the annotation
         LoincEntry loincEntry = loincTableView.getSelectionModel().getSelectedItem();
         LoincId loincCode = loincEntry.getLOINC_Number();
@@ -1276,6 +1476,7 @@ public class AnnotateTabController {
         flagForAnnotation.setSelected(false);
         annotationNoteField.clear();
         model.setSessionChanged(true);
+        loincTableEnableMultiSelection.setSelected(false);
 
         loinc2HpoAnnotationsTabController.refreshTable();
         createAnnotationSuccess.setFill(Color.GREEN);
@@ -1399,7 +1600,7 @@ public class AnnotateTabController {
 
     @FXML
     private void handleDragInTreeView(MouseEvent e) {
-        System.out.println("Drag event detected");
+        //System.out.println("Drag event detected");
         Dragboard db = treeView.startDragAndDrop(TransferMode.ANY);
         ClipboardContent content = new ClipboardContent();
         Object selectedItem = treeView.getSelectionModel().getSelectedItem().getValue();
@@ -1468,36 +1669,50 @@ public class AnnotateTabController {
             @Override
             protected void updateItem(String item, boolean empty){
                 super.updateItem(item, empty);
-                if(item != null && !empty) {
+                //if(item != null && !empty) {
+                if(!empty) {
                     setText(item);
                     try {
                         if(model.getLoincAnnotationMap().containsKey(new LoincId(item))) {
-                            logger.info("model contains " + item);
-                            logger.info("num of items in model " + model.getLoincAnnotationMap().size());
                             TableRow<LoincEntry> currentRow = getTableRow();
                             currentRow.setStyle("-fx-background-color: cyan");
-                            //setStyle("-fx-text-fill: red; -fx-font-weight: bold; -fx-background-color: lightblue");
-                        } else if (model.getUserCreatedLoincLists().get(LOINCWAITING4NEWHPO) != null
-                                && model.getUserCreatedLoincLists().get(LOINCWAITING4NEWHPO).contains(new LoincId(item))) {
-                            logger.info("model usercreated list for LoincWaiting4NewHPO # items: " +
-                                    model.getUserCreatedLoincLists().get(LOINCWAITING4NEWHPO).size());
+                        } else {
                             TableRow<LoincEntry> currentRow = getTableRow();
-                            currentRow.setStyle("-fx-background-color: deeppink");
-                        } else if (model.getUserCreatedLoincLists().get(LOINCUNABLE2ANNOTATE) != null
-                                && model.getUserCreatedLoincLists().get(LOINCUNABLE2ANNOTATE).contains(new LoincId(item))) {
-                            TableRow<LoincEntry> currentRow = getTableRow();
-                            currentRow.setStyle("-fx-background-color: lightcoral");
-                            //@TODO: change color of other groups. tip: allow user to pick a color
-                        } else{//for reasons I don't understand, this else block is critical to make it work!!!
-                            TableRow<LoincEntry> currentRow = getTableRow();
-                            currentRow.setStyle("");
+                            LoincId loincId = new LoincId(item);
+                            List<String> inList = model.getUserCreatedLoincLists().entrySet()
+                                    .stream()
+                                    .filter(entry -> entry.getValue().contains(loincId))
+                                    .map(entry -> entry.getKey())
+                                    .collect(Collectors.toList());
+                            if (!inList.isEmpty()) {
+                                List<Color> colors = inList.stream()
+                                    .map(l -> model.getUserCreatedLoincListsColor().get(l))
+                                    .filter(Objects::nonNull)
+                                    .map(Color::web)
+                                    .collect(Collectors.toList());
+                                if (colors.isEmpty()) {
+                                    currentRow.setStyle("");
+                                } else {
+                                    String backgroundColorValue = colors.get(0).toString(); //just use the first color
+                                    logger.trace(backgroundColorValue);
+                                    logger.trace(String.format("#%s", backgroundColorValue.substring(2,8).toUpperCase()));
+                                    currentRow.setStyle("-fx-background-color: " + String.format("#%s", backgroundColorValue.substring(2,8).toUpperCase()));
+                                    //Cannot use set background. It DOES NOT work!
+                                    //currentRow.setBackground(new Background(new BackgroundFill(colors.get(0), null, null)));
+                                }
+
+                            } else {
+                                currentRow.setStyle("");
+                            }
                         }
                     } catch (MalformedLoincCodeException e) {
                         //do nothing
+                        logger.error("should never happen:xdeide");
                     }
                 } else {
                     setText(null);
                     getTableRow().setStyle("");
+                    //logger.trace("changecolor:44444");
                 }
             }