+++ loinc2hpogui/src/main/java/org/monarchinitiative/loinc2hpo/model/Model.java
@@ -8,11 +8,16 @@ import org.hl7.fhir.dstu3.model.Observation;
 import org.monarchinitiative.loinc2hpo.Constants;
 import org.monarchinitiative.loinc2hpo.io.HpoOntologyParser;
 import org.monarchinitiative.loinc2hpo.loinc.*;
+import org.monarchinitiative.phenol.base.PhenolException;
 import org.monarchinitiative.phenol.formats.hpo.HpoOntology;
 import org.monarchinitiative.phenol.ontology.data.*;
 
 import java.io.*;
 import java.util.*;
+import java.util.List;
+import java.util.stream.Collectors;
+import javafx.scene.paint.Color;
+import org.semanticweb.owlapi.model.OWLOntologyCreationException;
 
 /**
  * Prototype model for LOINC to HPO Biocuration process.
@@ -50,6 +55,7 @@ public class Model {
     /** Key: a loinc code such as 10076-3; value: the corresponding TODO -- what link QnLoinc2HPOAnnotation object .*/
     public Map<LoincId,LOINC2HpoAnnotationImpl> loincAnnotationMap =new LinkedHashMap<>();
     private Map<String, Set<LoincId>> userCreatedLoincLists = new LinkedHashMap<>();
+    private Map<String, String> userCreatedLoincListsColor = new LinkedHashMap<>();
 
     private Map<LoincId, LoincEntry> loincEntryMap;
     private HashSet<LoincId> loincIds = new HashSet<>();
@@ -316,6 +322,12 @@ public class Model {
         this.userCreatedLoincLists.put(listName, list);
     }
 
+    public Map<String, String> getUserCreatedLoincListsColor() { return userCreatedLoincListsColor; }
+
+    public void addOrUpdateUserCreatedLoincListColor(String listName, String color) {
+        userCreatedLoincListsColor.put(listName, color);
+    }
+
     private void init() {
     }
 
@@ -325,12 +337,14 @@ public class Model {
             logger.error("Attempt to parse hp.obo file with null path to file");
             return;
         }
-        HpoOntologyParser parser = new HpoOntologyParser(pathToHpoOboFile);
+        HpoOntologyParser parser = new HpoOntologyParser(pathToHpoOwlFile);
         try {
             parser.parseOntology();
             this.ontology = parser.getOntology();
-        } catch (IOException e) {
+        } catch (PhenolException e) {
             logger.error("Could not parse HPO obo file at "+pathToHpoOboFile);
+        } catch (OWLOntologyCreationException e) {
+            logger.error("Could not parge HPO owl file at " + pathToHpoOwlFile);
         }
         termmap=parser.getTermMap();
         termmap2=parser.getTermMap2();
@@ -376,6 +390,14 @@ public class Model {
             if (pathToHpGitRepo != null) {
                 bw.write(String.format("hp-repo:%s\n", pathToHpGitRepo));
             }
+            if (!userCreatedLoincListsColor.isEmpty()) {
+                bw.write("loinc-list-color:");
+                List<String> list_color_pair = userCreatedLoincListsColor.entrySet().stream()
+                        .map(e -> e.getKey() + "," + e.getValue())
+                        .collect(Collectors.toList());
+                bw.write(String.join("|", list_color_pair));
+                bw.write("\n");
+            }
             bw.close();
         } catch (IOException e) {
             e.printStackTrace();
@@ -408,6 +430,13 @@ public class Model {
                 else if (key.equals("autosave to")) this.pathToAutoSavedFolder = value;
                 else if (key.equals("last session")) this.pathToLastSession = value;
                 else if (key.equals("hp-repo")) this.pathToHpGitRepo = value;
+                else if (key.equals("loinc-list-color")) {
+                    String[] entries = value.split("\\|");
+                    Arrays.stream(entries)
+                            .map(e -> e.split(",")) //has to be two elements
+                            .forEach(e -> userCreatedLoincListsColor.put(e[0], e[1]));
+                    logger.trace("color for LOINC lists is set from settings");
+                }
             }
             br.close();
         } catch (IOException e) {
@@ -433,6 +462,21 @@ public class Model {
     public void setFhirServers(List<String> fhirServers) {
         this.fhirServers = fhirServers;
     }
+
+    public List<String> defaultColorList() {
+        return Arrays.asList(
+                Color.CYAN.toString(),
+                Color.MAGENTA.toString(),
+                Color.PINK.toString(),
+                Color.LIGHTBLUE.toString(),
+                Color.ORANGE.toString(),
+                Color.CHOCOLATE.toString()
+                );
+                //.stream()
+                //.map(color -> "#" + color.substring(2,8).toUpperCase())
+                //.collect(Collectors.toList());
+    }
+